backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
torch
torchvision
timm
pillow
numpy
pydantic
pydantic-settings
rasterio
opencv-python-headless
python-dotenv
aiofiles
